# python_apps/workflow-engine-poc/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Unified Workflow Execution Engine",
        "description": "A clean, agnostic workflow execution engine PoC",
        "version": "1.0.0-poc"
    },
    "paths": {
...
        "/executions/{execution_id}": {
            "get": {
                "tags": [
                    "executions"
                ],
                "summary": "Get Execution Status",
                "description": "Get the status of a workflow execution",
                "operationId": "get_execution_status_executions__execution_id__get",
                "parameters": [
                    {
                        "name": "execution_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Execution Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/executions/{execution_id}/results": {
            "get": {
                "tags": [
                    "executions"
                ],
                "summary": "Get Execution Results",
                "description": "Get the detailed results of a workflow execution",
                "operationId": "get_execution_results_executions__execution_id__results_get",
                "parameters": [
                    {
                        "name": "execution_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Execution Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/executions/": {
            "get": {
                "tags": [
                    "executions"
                ],
                "summary": "Get Execution Analytics",
                "description": "Get execution analytics and history",
                "operationId": "get_execution_analytics_executions__get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Offset"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Status"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/executions/{workflow_id}/execute": {
            "post": {
                "tags": [
                    "executions"
                ],
                "summary": "Create Execution",
                "description": "Create a new workflow execution using SQLModel and return ORM entity",
                "operationId": "create_execution_executions__workflow_id__execute_post",
                "parameters": [
                    {
                        "name": "workflow_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Workflow Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "additionalProperties": true,
                                "title": "Execution Data"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkflowExecutionRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/executions/sqlmodel/{execution_id}": {
            "get": {
                "tags": [
                    "executions"
                ],
                "summary": "Get Execution Sqlmodel",
                "description": "Get execution details using SQLModel and return ORM entity",
                "operationId": "get_execution_sqlmodel_executions_sqlmodel__execution_id__get",
                "parameters": [
                    {
                        "name": "execution_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Execution Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkflowExecutionRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/executions/sqlmodel/": {
            "get": {
                "tags": [
                    "executions"
                ],
                "summary": "List Executions Sqlmodel",
                "description": "List executions using SQLModel and return ORM entities",
                "operationId": "list_executions_sqlmodel_executions_sqlmodel__get",
                "parameters": [
                    {
                        "name": "workflow_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Workflow Id"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Status"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Offset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WorkflowExecutionRead"
                                    },
                                    "title": "Response List Executions Sqlmodel Executions Sqlmodel  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/analytics/executions": {
            "get": {
                "tags": [
                    "monitoring"
                ],
                "summary": "Get Execution Analytics",
                "description": "Get execution analytics and history",
                "operationId": "get_execution_analytics_analytics_executions_get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Offset"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Status"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ExecutionStatus": {
                "type": "string",
                "enum": [
                    "pending",
                    "running",
                    "completed",
                    "failed",
                    "cancelled",
                    "timeout"
                ],
                "title": "ExecutionStatus",
                "description": "Execution status enumeration"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
            "WorkflowExecutionRead": {
                "properties": {
                    "workflow_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Workflow Id",
                        "description": "ID of the workflow being executed"
                    },
                    "user_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 255
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User Id",
                        "description": "User who triggered the execution"
                    },
                    "session_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 255
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Session Id",
                        "description": "Session ID for tracking"
                    },
                    "organization_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 255
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Organization Id",
                        "description": "Organization ID"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ExecutionStatus",
                        "description": "Current execution status",
                        "default": "pending"
                    },
                    "input_data": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Input Data",
                        "description": "Initial input data for the workflow"
                    },
                    "output_data": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Output Data",
                        "description": "Final output data from the workflow"
                    },
                    "step_io_data": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Step Io Data",
                        "description": "Input/output data for all steps"
                    },
                    "execution_metadata": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Execution Metadata",
                        "description": "Execution metadata and analytics"
                    },
                    "error_message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error Message",
                        "description": "Error message if execution failed"
                    },
                    "started_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Started At",
                        "description": "When execution started"
                    },
                    "completed_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Completed At",
                        "description": "When execution completed"
                    },
                    "execution_time_seconds": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Execution Time Seconds",
                        "description": "Total execution time in seconds"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Updated At"
                    },
                    "total_steps": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Total Steps",
                        "description": "Total number of steps"
                    },
                    "completed_steps": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Completed Steps",
                        "description": "Number of completed steps"
                    },
                    "failed_steps": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Failed Steps",
                        "description": "Number of failed steps"
                    }
                },
                "type": "object",
                "required": [
                    "workflow_id",
                    "id",
                    "created_at"
                ],
                "title": "WorkflowExecutionRead",
                "description": "Schema for reading workflow execution data"
            },
...
        }
    }
}